requirements.txt
google-genai
virtualenv
django
python-dotenv